Using the compose avatar in settings
