+++ app/src/main/java/org/thoughtcrime/securesms/glide/PlaceholderAvatarFetcher.kt
@@ -7,14 +7,14 @@ import com.bumptech.glide.load.DataSource
 import com.bumptech.glide.load.data.DataFetcher
 import org.session.libsession.avatars.PlaceholderAvatarPhoto
 import org.session.libsignal.utilities.Log
-import org.thoughtcrime.securesms.util.AvatarPlaceholderGenerator
+import org.thoughtcrime.securesms.util.AvatarUtils
 
 class PlaceholderAvatarFetcher(private val context: Context,
                                private val photo: PlaceholderAvatarPhoto): DataFetcher<BitmapDrawable> {
 
     override fun loadData(priority: Priority,callback: DataFetcher.DataCallback<in BitmapDrawable>) {
         try {
-            val avatar = AvatarPlaceholderGenerator.generate(context, 128, photo.hashString, photo.displayName)
+            val avatar = AvatarUtils.generateTextBitmap(context, 128, photo.hashString, photo.displayName)
             callback.onDataReady(avatar)
         } catch (e: Exception) {
             Log.e("Loki", "Error in fetching avatar")
