+++ app/src/main/java/org/thoughtcrime/securesms/ui/components/Avatar.kt
@@ -169,7 +169,7 @@ fun PreviewAvatarElement(){
         AvatarElement(
             size = LocalDimensions.current.iconLarge,
             data = AvatarUIData(
-                name = "AT",
+                name = "TO",
                 color = primaryGreen,
                 contactPhoto = null
             )
