+++ app/src/main/res/layout/activity_settings.xml
@@ -21,10 +21,10 @@
             android:orientation="vertical"
             android:gravity="center_horizontal">
 
-            <org.thoughtcrime.securesms.components.ProfilePictureView
-                android:id="@+id/profilePictureView"
-                android:layout_width="@dimen/large_profile_picture_size"
-                android:layout_height="@dimen/large_profile_picture_size"
+            <androidx.compose.ui.platform.ComposeView
+                android:id="@+id/userAvatar"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
                 android:layout_marginTop="@dimen/medium_spacing"
                 android:contentDescription="@string/AccessibilityId_profilePicture" />
