+++ app/src/main/res/values/colors.xml
@@ -13,16 +13,6 @@
     <color name="call_action_button_highlighted">#FFFFFF</color>
     <color name="call_action_foreground_highlighted">#171717</color>
 
-    <array name="profile_picture_placeholder_colors">
-        <item>@color/accent_green</item>
-        <item>@color/accent_blue</item>
-        <item>@color/accent_yellow</item>
-        <item>@color/accent_pink</item>
-        <item>@color/accent_purple</item>
-        <item>@color/accent_orange</item>
-        <item>@color/accent_red</item>
-    </array>
-
     <color name="loki_darkest_gray">#0a0a0a</color>
 
     <color name="white">#ffffffff</color>
