+++ libsession/src/main/res/values/colors.xml
@@ -9,13 +9,6 @@
     <color name="compose_text_view_background">#141414</color>
     <color name="paths_building">#FFCE3A</color>
 
-    <array name="profile_picture_placeholder_colors">
-        <item>#5ff8b0</item>
-        <item>#26cdb9</item>
-        <item>#f3c615</item>
-        <item>#fcac5a</item>
-    </array>
-
     <color name="loki_darkest_gray">#0a0a0a</color>
 
     <color name="signal_primary">@color/accent_green</color>
